requirements.txt
pillow==10.0.0
numpy==1.24.3
tk==0.1.0
scipy==1.11.3